Remove jruby from travis
As of 7/6/2019, the build was passing in jruby (9.2.7.0), however, the
build was a lot slower than the MRI's.  Also, because jruby doesn't
actually support continuations, but backs its fibers with actual
threads, it's not clear to me at this point whether shifty is a good
fit for jruby (i.e. lots of workers means lots of fibers, means _lots_
of threads).

I did, however, drop in builds for MRI 2.4 and 2.3.